Updated the Save/Load System
Added "update save" functionality to the save system, now actually
serving a purpose in saving your game as you play.
as of now, the player needs to be removed from the gameboard before
Save.updateSave is called, in order to keep the current player's
Entity.toString from clusterfucking up the save.

In the future, I will add tracking equipment, inventory, and unspent SP.