Work, Work, Work, Work, Work, Hello Boys, I Missed You!
added the rl4j and slick libraries into the folder, such that the whole
project can be used as a single, self-referential (and therefore
portable?) eclipse project, for the sake of simplicity.

Fixed glitches in the save system, and added a "blocking" tag to tiles.
The point of this tag is to avoid the old problem with millions of
useless, memory eating "Wall" objects.

Altered the colors used in the minimap slightly

Added HP and MP tracking to saves, made saving accessible at any time
(press P).

Fixed a glitch where the HP and MP bars would accelerate to way past
maximum values on extended wait times (during large calculations)