Doing things.
Sexy Things.

Altered EntityTypeAction to run based off of isAssignable, as oppose to
try-catch casting ugliness.

Doors will no longer close on top of entities, deleting themselves in
the process.

in Entities, isInteractive() renamed to makesActions to avoid confusion
down the line.

Began shift towards static-variable based keybindings, to allow for
rebindable keys in the future

changed the lootstate to a custom state, removed need for
ItemGridInteractionElement in the process. (as yet unfinished
polishing).