Fucking modloading, mate
forgot to capitalize a T "wallTiles.png" on in the core modpack,
meaning it worked when DeadReckoningMods/src was on the build path, but
not when it wasn't. Now it should (and does, in my brief period of
testing) function with it off the build path as it should have.

this took me half an hour to figure out.

I hate debugging.