Big changes are coming, kid.
replaced Will(lego6245)'s error handling system with a better
alternative that allows for catching runtime errors. To do this, the I
had to copy some of the code out of the existing slick files, because
calling super() was not an option (needed to bypass the existing error
handling). As a result, we now have DeadReckoningContainer, and
FileSaveLogSystem

I also began moving some very reusable biome code (Pathmaking) into a
new class, BiomeUtils.

began work on an entityID system, but y'knowm shit sux. very unstable,
as all of the transitional code is not in place yet.

Continued implementing missign theif skills. not done at all. UNSTABLE.

added slick src files for easy reference.